* stable-2.16
  Fix lines with more than 80 characters
  Add more detach/attach sequence tests
  Allow disk attachment to diskless instances
  Improve tests for attaching disks
  Use only string value in error message
  Add entries describing new gnt-cluster params to manpage
  QA: Add ssh-key-type and -bits tests
  QA: Extend AssertCommand to allow not forwarding the agent
  Remove default limit on diffs in cfgupgrade tests
  QA: Downgrade the cluster key type in 2.16
  Fix typo
  Fail early for invalid key type and size combinations
  Handle SSH key changes in upgrades and downgrades
  Allow SSH key property changes
  Use the SSH key parameters when generating keys
  Do not generate the ganeti_pub_keys file with --no-ssh-init
  Add querying of ssh-related config values
  Add modify_ssh_setup to queryable config params
  Add helper function for querying cluster properties
  Show info about new params in gnt-cluster info
  Add the SSH key type and length to the config, and set them
  Change SSH key types to a proper Haskell sum type
  Add the SSH key options
  Mention disabling of '--no-node-setup' in NEWS file
  Show 'modify ssh setup' in cluster info
  Disable --no-node-setup
  Make 'modify ssh setup' queryable
  Fix RPC signature of NodeVerify
  Use ssconf for SSH ports in NodeVerify
* stable-2.15
  Document the decission why optimisation is turned off
  Don't keep input for error messages
  Use dict.copy instead of deepcopy
  Use bulk-adding of keys in renew-crypto
  Make NodeSshKeyAdd use its *Bulk companion
  Unit test bulk-adding normal nodes
  Unit test for bulk-adding pot. master candidates
  Introduce bulk-adding of SSH keys
  Pause watcher during performance QA
  Send answers strictly
  Store keys as ByteStrings
  Encode UUIDs as ByteStrings
  Prefer the UuidObject type class over specific functions
  Assign the variables before use (bugfix for dee6adb9)
  Extend QA to detect autopromotion errors
  Handle SSH key distribution on auto promotion
  Do not remove authorized key of node itself
  Fix indentation
  Support force option for deactivate disks on RAPI
* stable-2.14
  Fix faulty iallocator type check
  Improve cfgupgrade output in case of errors
* stable-2.13
  Extend timeout for gnt-cluster renew-crypto
  Reduce flakyness of GetCmdline test on slow machines
  Remove duplicated words
* stable-2.12
  Revert "Also consider connection time out a network error"
  Clone lists before modifying
  Make lockConfig call retryable
  Return the correct error code in the post-upgrade script
  Make openssl refrain from DH altogether
  Fix upgrades of instances with missing creation time
* stable-2.11
  (no changes)
* stable-2.10
  Remove -X from hspace man page
  Make htools tolerate missing "dtotal" and "dfree" on luxi

Conflicts:
	NEWS
	lib/cli_opts.py
	lib/objects.py
	src/Ganeti/Config.hs
	src/Ganeti/DataCollectors.hs
	src/Ganeti/Monitoring/Server.hs
	src/Ganeti/Objects.hs
	src/Ganeti/Objects/Disk.hs
	src/Ganeti/Objects/Instance.hs
	src/Ganeti/Query/Group.hs
	src/Ganeti/Query/Server.hs
	src/Ganeti/WConfd/ConfigModifications.hs
	src/Ganeti/WConfd/ConfigVerify.hs
	test/hs/Test/Ganeti/Objects.hs
	test/py/cfgupgrade_unittest.py

Resolution:
        NEWS
          take both changes
        lib/cli_opts.py
          take both changes
        lib/objects.py
          take both changes
        src/Ganeti/Config.hs
          keep the ByteString changes, but Control.Monad from 2.17
        src/Ganeti/DataCollectors.hs
          take both changes
        src/Ganeti/Monitoring/Server.hs
          fix imports
        src/Ganeti/Objects.hs
          take both changes
        src/Ganeti/Objects/Disk.hs
          take both changes
        src/Ganeti/Objects/Instance.hs
          fix imports
          keep 2.17 changes
        src/Ganeti/Query/Group.hs
          keep field definition for hv_state and disk_state, but use
          uuidOf instead of groupUuid
        src/Ganeti/Query/Server.hs
          take both changes
        src/Ganeti/WConfd/ConfigModifications.hs
          fix imports
        src/Ganeti/WConfd/ConfigVerify.hs
          fix imports
        test/hs/Test/Ganeti/Objects.hs
          fix imports
          take both changes
        test/py/cfgupgrade_unittest.py
          take both changes

+++ lib/client/gnt_cluster.py
@@ -301,6 +301,14 @@ def InitCluster(opts, args):
   else:
     enabled_user_shutdown = False
 
+  if opts.ssh_key_type:
+    ssh_key_type = opts.ssh_key_type
+  else:
+    ssh_key_type = constants.SSH_DEFAULT_KEY_TYPE
+
+  ssh_key_bits = ssh.DetermineKeyBits(ssh_key_type, opts.ssh_key_bits, None,
+                                      None)
+
   bootstrap.InitCluster(cluster_name=args[0],
                         secondary_ip=opts.secondary_ip,
                         vg_name=vg_name,
@@ -335,6 +343,8 @@ def InitCluster(opts, args):
                         zeroing_image=zeroing_image,
                         compression_tools=compression_tools,
                         enabled_user_shutdown=enabled_user_shutdown,
+                        ssh_key_type=ssh_key_type,
+                        ssh_key_bits=ssh_key_bits,
                         )
   op = opcodes.OpClusterPostInit()
   SubmitOpCode(op, opts=opts)
@@ -614,6 +624,9 @@ def ShowClusterConfig(opts, args):
       ("zeroing image", result["zeroing_image"]),
       ("compression tools", result["compression_tools"]),
       ("enabled user shutdown", result["enabled_user_shutdown"]),
+      ("modify ssh setup", result["modify_ssh_setup"]),
+      ("ssh_key_type", result["ssh_key_type"]),
+      ("ssh_key_bits", result["ssh_key_bits"]),
       ]),
 
     ("Default node parameters",
@@ -966,11 +979,12 @@ def _ReadAndVerifyCert(cert_filename, verify_private_key=False):
   return pem
 
 
+# pylint: disable=R0913
 def _RenewCrypto(new_cluster_cert, new_rapi_cert, # pylint: disable=R0911
                  rapi_cert_filename, new_spice_cert, spice_cert_filename,
                  spice_cacert_filename, new_confd_hmac_key, new_cds,
                  cds_filename, force, new_node_cert, new_ssh_keys,
-                 verbose, debug):
+                 ssh_key_type, ssh_key_bits, verbose, debug):
   """Renews cluster certificates, keys and secrets.
 
   @type new_cluster_cert: bool
@@ -998,10 +1012,14 @@ def _RenewCrypto(new_cluster_cert, new_rapi_cert, # pylint: disable=R0911
   @param new_node_cert: Whether to generate new node certificates
   @type new_ssh_keys: bool
   @param new_ssh_keys: Whether to generate new node SSH keys
+  @type ssh_key_type: One of L{constants.SSHK_ALL}
+  @param ssh_key_type: The type of SSH key to be generated
+  @type ssh_key_bits: int
+  @param ssh_key_bits: The length of the key to be generated
   @type verbose: boolean
-  @param verbose: show verbose output
+  @param verbose: Show verbose output
   @type debug: boolean
-  @param debug: show debug output
+  @param debug: Show debug output
 
   """
   ToStdout("Updating certificates now. Running \"gnt-cluster verify\" "
@@ -1182,7 +1200,9 @@ def _RenewCrypto(new_cluster_cert, new_rapi_cert, # pylint: disable=R0911
     cl = GetClient()
     renew_op = opcodes.OpClusterRenewCrypto(
         node_certificates=new_node_cert or new_cluster_cert,
-        ssh_keys=new_ssh_keys)
+        renew_ssh_keys=new_ssh_keys,
+        ssh_key_type=ssh_key_type,
+        ssh_key_bits=ssh_key_bits)
     SubmitOpCode(renew_op, cl=cl)
 
   ToStdout("All requested certificates and keys have been replaced."
@@ -1199,17 +1219,24 @@ def _BuildGanetiPubKeys(options, pub_key_file=pathutils.SSH_PUB_KEYS, cl=None,
   """Recreates the 'ganeti_pub_key' file by polling all nodes.
 
   """
+
+  if not cl:
+    cl = GetClient()
+
+  (cluster_name, master_node, modify_ssh_setup, ssh_key_type) = \
+    cl.QueryConfigValues(["cluster_name", "master_node", "modify_ssh_setup",
+                          "ssh_key_type"])
+
+  # In case Ganeti is not supposed to modify the SSH setup, simply exit and do
+  # not update this file.
+  if not modify_ssh_setup:
+    return
+
   if os.path.exists(pub_key_file):
     utils.CreateBackup(pub_key_file)
     utils.RemoveFile(pub_key_file)
 
   ssh.ClearPubKeyFile(pub_key_file)
-
-  if not cl:
-    cl = GetClient()
-
-  (cluster_name, master_node) = \
-    cl.QueryConfigValues(["cluster_name", "master_node"])
 
   online_nodes = get_online_nodes_fn([], cl=cl)
   ssh_ports = get_nodes_ssh_ports_fn(online_nodes + [master_node], cl)
@@ -1223,7 +1250,7 @@ def _BuildGanetiPubKeys(options, pub_key_file=pathutils.SSH_PUB_KEYS, cl=None,
 
   _, pub_key_filename, _ = \
     ssh.GetUserFiles(constants.SSH_LOGIN_USER, mkdir=False, dircheck=False,
-                     kind=constants.SSHK_DSA, _homedir_fn=homedir_fn)
+                     kind=ssh_key_type, _homedir_fn=homedir_fn)
 
   # get the key file of the master node
   pub_key = utils.ReadFile(pub_key_filename)
@@ -1257,6 +1284,8 @@ def RenewCrypto(opts, args):
                       opts.force,
                       opts.new_node_cert,
                       opts.new_ssh_keys,
+                      opts.ssh_key_type,
+                      opts.ssh_key_bits,
                       opts.verbose,
                       opts.debug > 0)
 
@@ -2432,7 +2461,7 @@ commands = {
      HV_STATE_OPT, DISK_STATE_OPT, ENABLED_DISK_TEMPLATES_OPT,
      IPOLICY_STD_SPECS_OPT, GLOBAL_GLUSTER_FILEDIR_OPT, INSTALL_IMAGE_OPT,
      ZEROING_IMAGE_OPT, COMPRESSION_TOOLS_OPT,
-     ENABLED_USER_SHUTDOWN_OPT,
+     ENABLED_USER_SHUTDOWN_OPT, SSH_KEY_BITS_OPT, SSH_KEY_TYPE_OPT,
      ]
      + INSTANCE_POLICY_OPTS + SPLIT_ISPECS_OPTS,
     "[opts...] <cluster_name>", "Initialises a new cluster configuration"),
@@ -2534,7 +2563,7 @@ commands = {
      NEW_CLUSTER_DOMAIN_SECRET_OPT, CLUSTER_DOMAIN_SECRET_OPT,
      NEW_SPICE_CERT_OPT, SPICE_CERT_OPT, SPICE_CACERT_OPT,
      NEW_NODE_CERT_OPT, NEW_SSH_KEY_OPT, NOSSH_KEYCHECK_OPT,
-     VERBOSE_OPT],
+     VERBOSE_OPT, SSH_KEY_BITS_OPT, SSH_KEY_TYPE_OPT],
     "[opts...]",
     "Renews cluster certificates, keys and secrets"),
   "epo": (
